backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-multipart==0.0.9
transformers==4.40.1
torch==2.2.2
accelerate==0.30.1
soundfile==0.12.1
librosa==0.10.2
pydub==0.25.1
pyannote.audio==3.2.0
nemo-toolkit[asr]==1.23.0
vllm==0.4.2
huggingface-hub==0.21.4
uvloop==0.19.0
orjson==3.10.3
redis==5.0.4
aiofiles==23.2.1
psutil==5.9.8
gputil==1.4.0
